created LibraryTestController for testing badges from model
