+++ src/main/java/com/greenfox/lvlup/controller/LibraryTestController.java
@@ -1,0 +1,27 @@
+package com.greenfox.lvlup.controller;
+
+import com.greenfox.lvlup.model.entityTestingDto.BadgeDTO;
+import com.greenfox.lvlup.service.BadgeService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+public class LibraryTestController {
+    private BadgeService badgeService;
+
+    @Autowired
+    public LibraryTestController(BadgeService badgeService) {
+        this.badgeService = badgeService;
+    }
+
+    @GetMapping("/library")
+    public List<BadgeDTO> getAllDtos() {
+        return badgeService.getDTOListFromBadge();
+    }
+
+
+}
+
